# docs/swagger.yaml
definitions:
  controller.CreateConfessionRequest:
    properties:
      anon:
        type: boolean
      text:
        type: string
      title:
        type: string
    required:
    - text
    - title
    type: object
  controller.UpdateConfessionRequest:
    properties:
      anon:
        type: boolean
      text:
        type: string
      title:
        type: string
    type: object
  models.Confession:
    properties:
      anon:
        type: boolean
      created_at:
        type: string
      guest_uuid:
        type: string
      id:
        type: integer
      text:
        type: string
      title:
        maxLength: 100
        minLength: 5
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    required:
    - text
    - title
    type: object
...
info:
  contact: {}
  description: API Server for Confessly Application
  title: Confessly API
  version: "1.0"
paths:
...
  /admin/confessions/{id}:
    delete:
      parameters:
      - description: Confession ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удаление конфесии по ID (только для администраторов)
      tags:
      - admin
...
  /confessions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Confession'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение всех конфесий
      tags:
      - confession
    post:
      consumes:
      - application/json
      parameters:
      - description: Confession object
        in: body
        name: confession
        required: true
        schema:
          $ref: '#/definitions/controller.CreateConfessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создание конфесии
      tags:
      - confession
  /confessions/{id}:
    delete:
      parameters:
      - description: Confession ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удаление конфесии
      tags:
      - confession
    get:
      parameters:
      - description: Confession ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Confession'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение конфесии по ID
      tags:
      - confession
    put:
      consumes:
      - application/json
      parameters:
      - description: Confession ID
        in: path
        name: id
        required: true
        type: integer
      - description: Confession object
        in: body
        name: confession
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateConfessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление конфесии
      tags:
      - confession
  /confessions/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Confession'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Поиск конфесий по названию
      tags:
      - confession
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"